main_app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
google-adk
google-generativeai
httpx
python-dotenv
crawl4ai
pydantic

# AIzaSyDtyopWRWXIYnYkBaOrlnb4t-h3RABeOds